pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "slow: marks tests as very slow (deselect with '-m \"not slow\"')",
]
